backend/requirements.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
Pillow
requests
gunicorn
whitenoise
dj-database-url
psycopg2-binary
